% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_textClust.R
\name{DSC_textClust}
\alias{DSC_textClust}
\title{textClust}
\usage{
DSC_textClust(r = 0.4, lambda = 0.1, tgap = 1000, nmin = 1,
  nmax = 1, k = NA_integer_, h = NA_real_, verbose = F,
  termFading = T, stopword = stopwords(language = "en", source =
  "stopwords-iso"), linkage = "complete", weightedReclustering = TRUE,
  minWeight = 2, textCol = 1, timeCol = NA_integer_,
  groupByCol = NA_integer_, parentTextCol = NA_integer_,
  parentTimeCol = NA_integer_, timeFormat = "\%Y-\%m-\%d \%H:\%M:\%S",
  timePrecision = "days", fadeNaturalTime = FALSE)
}
\arguments{
\item{r}{distance threshold to merge two micro-clusters}

\item{lambda}{fading constant}

\item{tgap}{time between outlier removal}

\item{nmin}{min number of ngrams to use}

\item{nmax}{max number of ngrams to use}

\item{k}{number of clusters for macro-clustering}

\item{h}{height to determine number of clusters for macro-clustering}

\item{verbose}{logical whether to be more verbose}

\item{termFading}{logical whether individual terms should also be faded}

\item{stopword}{chracter vector of stopwords to remove}

\item{linkage}{method for hierarchical clustering}

\item{weightedReclustering}{logical whether reclustering should consider cluster weights}

\item{minWeight}{minimum weight of micro clusters to be used for reclustering}

\item{textCol}{index of column that contains the text which should be clustered}

\item{timeCol}{index of column that contains timestamps}

\item{groupByCol}{index of column that groups text into conversations (i.e. multiple texts into the same document)}

\item{parentTextCol}{index of column that contains the text of the parent when using groupByCol}

\item{parentTimeCol}{index of column that contains the time of the parent when using groupByCol}

\item{timeFormat}{string formatting of time Column}

\item{timePrecision}{Precision of fading, either seconds, minutes, hours or days}

\item{fadeNaturalTime}{Logical whether Natural Time or Number of observations should be used for fading}
}
\value{
micro clusters
}
\description{
Stream Clustering algorithm which clusters text data.
}
\examples{
data = data.frame(text=sample(c("Main Topic", "Similar Topic", "Something Different"), size=1000, replace=T),stringsAsFactors=F)
stream = DSD_Memory(data)  ## Alternatively: stream = DSD_ReadCSV("file.txt", sep = "\\t", comment.char="", quote="")
algorithm = DSC_textClust(r=.4, lambda=0.1, tgap=100, nmin=1, nmax=2, k=3, stopword=c(), minWeight=3, textCol=1)
update(algorithm, stream, n=1000)

}
\references{
Matthias Carnein, Dennis Assenmacher, Heike Trautmann (2017)"Stream Clustering of Chat Messages with Applications to Twitch Streams". In: Advances in Conceptual Modeling. ER 2017. Lecture Notes in Computer Science, vol 10651. Springer, Cham
}
\author{
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
