% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_textClust.R
\name{DSD_textClust}
\alias{DSD_textClust}
\title{DSD_textClust}
\usage{
DSD_textClust(streams, labels = NULL, time = T)
}
\arguments{
\item{streams}{vector of stream file names}

\item{labels}{vector that specifies how the streams should be labelled.}

\item{time}{specifies whether the stream should be ordered by time}
}
\description{
StreamClustering data source
}
\author{
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
