% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate Clustering}
\usage{
evaluate(dsc, dsd, measure, n = 100, type = c("auto", "micro",
  "macro"), assign = "micro", assignmentMethod = c("auto", "model",
  "nn"), noise = c("class", "exclude"), ...)
}
\description{
Adaption of evaluation method from the stream pacakge. Adapted for textual data.
}
