% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_cluster}
\alias{evaluate_cluster}
\title{Evaluate Clustering Using Prequential Evaluation}
\usage{
evaluate_cluster(dsc, dsd, measure, n = 1000, type = c("auto", "micro",
  "macro"), assign = "micro", assignmentMethod = c("auto", "model",
  "nn"), horizon = 100, verbose = FALSE, noise = c("class",
  "exclude"), limit = Inf, ...)
}
\description{
Adaption of evaluate_cluster method from the stream pacakge.
}
