% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_textClust.R
\name{get_assignment.DSC_textClust}
\alias{get_assignment.DSC_textClust}
\title{Get Assignment for DSC_textClust}
\usage{
get_assignment.DSC_textClust(dsc, points, ...)
}
\arguments{
\item{dsc}{object of class DSC_textClust}

\item{points}{matrix with the messages to assign}
}
\value{
Indicies of the closest micro-clusters
}
\description{
Returns the index of the closest micro-cluster for every entry in points using the cosine similary of the tf-idf vectors.
}
\examples{
stream = DSD_ReadCSV("file.txt", sep = "\\t", comment.char="", quote="")
algorithm = DSC_textClust(lambda=0.001, r=0.2, tgap=100, verbose=T)
update(algorithm, stream, n=100)

data = read.table("file.txt", header=F, sep="\\t", quote="", comment.char = "", stringsAsFactors =F)
get_assignment(algorithm, data)

}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
