% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSD_textClust.R
\name{get_points.DSD_textClust}
\alias{get_points.DSD_textClust}
\title{Get Points from textClust source}
\usage{
get_points.DSD_textClust(x, n = 1, cluster = FALSE, ...)
}
\arguments{
\item{x}{\code{DSD_textClust} object to read the points from}

\item{n}{number of points to return. default: 1}

\item{cluster}{logical whether to attach cluster assignment as an attribute}
}
\description{
Return \code{n} points from a textClust data source.
}
\author{
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
