% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_textClust.R
\name{plot.DSC_textClust}
\alias{plot.DSC_textClust}
\title{Plot DSC_textClust}
\usage{
\method{plot}{DSC_textClust}(dsc, numRepresentatives = 1, type = "micro",
  dendrogram = F, merge = F, interactive = T)
}
\arguments{
\item{dsc}{object of class DSC_textClust}

\item{numRepresentatives}{number of words to plot per micro cluster (default 1)}

\item{type}{string indicating whether to plot micro or macro cluster}

\item{dendrogram}{logical whether to plot dendrogram or use multidimensional scaling to plot representatives}

\item{merge}{logical whether to merge all clusters into one}

\item{interactive}{use interactive plots with the zoom package}
}
\description{
Plot function for object of type DSC_textClust. Uses Multidimensional Scaling (MDS) to plot the words with the heighest weight for each micro-cluster
}
\examples{
stream = DSD_ReadCSV("file.txt", sep = "\\t", comment.char="", quote="")
algorithm = DSC_textClust(lambda=0.001, r=0.2, tgap=100, verbose=T)
update(algorithm, stream, n=100)
plot(algorithm, numRepresentatives = 2)

}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
