% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_textClust.R
\name{print.DSC_textClust}
\alias{print.DSC_textClust}
\title{Print DSC_textClust}
\usage{
\method{print}{DSC_textClust}(x, ...)
}
\arguments{
\item{...}{optional arguments are ignored}

\item{dsc}{object of class DSC_textClust}
}
\description{
Print an object of type DSC_textClust.
}
\examples{
algorithm = DSC_textClust(lambda=0.001, r=0.2, tgap=100, verbose=T)
print(algorithm)

}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
