% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust-package.R
\docType{package}
\name{textClust}
\alias{textClust}
\alias{textClust-package}
\title{Topic Discovery in Texts using Stream Clustering}
\description{
Topic Discovery in Texts using Stream Clustering
}
\references{
Matthias Carnein, Dennis Assenmacher, Heike Trautmann (2017)"Stream Clustering of Chat Messages with Applications to Twitch Streams". In: Advances in Conceptual Modeling. ER 2017. Lecture Notes in Computer Science, vol 10651. Springer, Cham
}
\author{
Matthias Carnein \email{Matthias.Carnein@uni-muenster.de}
Dennis Assenmacher \email{Dennis.assenmacher@wi.uni-muenster.de}
}
