% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_get_microweights}
\alias{textClust_get_microweights}
\title{get_macroclusters}
\value{
list of macro-clusters
}
\description{
Getter for weights of macro-clusters of textClust_R
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
