% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\docType{class}
\name{textClust_R-class}
\alias{textClust_R-class}
\alias{textClust_R}
\title{Reference Class textClust}
\description{
Maintains the current clustering and implements the stream package interfaces.
}
\section{Fields}{

\describe{
\item{\code{textClust_C}}{Exposed C++ class}

\item{\code{nmin}}{min number of ngrams to use}

\item{\code{nmax}}{max number of ngrams to use}

\item{\code{k}}{number of clusters}

\item{\code{h}}{height for hierarchical clustering}

\item{\code{upToDate}}{logical whether macro-clusters are up to date or need to be recomputed}

\item{\code{microMacroAssignment}}{assignment vector associating micro-clusters to a maco-cluster}

\item{\code{hc}}{hclust object}

\item{\code{throughput}}{stores the throughput of messages}

\item{\code{termFading}}{logical whether individual terms should be faded}
}}


\author{
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
