% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_calculateIDF}
\alias{textClust_R_calculateIDF}
\title{Calculate Inverted Document Frequency}
\value{
Inverted Document Frequency as a vector
}
\description{
Calculates the Inverted Document Frequency, i.e. the popularity of words accross the entire corpus.
}
\author{
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
