% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_cluster}
\alias{textClust_R_cluster}
\title{Clustering procedure of textClust}
\arguments{
\item{newdata}{matrix of text in a single column}
}
\description{
Main clustering procedure of textClust
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
