% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_countTerms}
\alias{textClust_R_countTerms}
\title{count Terms}
\description{
Counts the number of terms in the micro or macro clusters
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
