% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_deserialize}
\alias{textClust_R_deserialize}
\title{Deserialize DSC_textClust_R object}
\description{
Imports a serialized R-List back into an C++ Object
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
