% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_dist}
\alias{textClust_R_dist}
\title{Pairwise distances between micro-cluster}
\value{
distance matrix of class \code{dist}
}
\description{
Returns a pairwise distance matrix between all micro-clusters .
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
