% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_get_assignment}
\alias{textClust_R_get_assignment}
\title{get_assignment}
\description{
Returns the cluster assignment of text to a respective micro-clusters
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
