% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_get_macroclusters}
\alias{textClust_R_get_macroclusters}
\title{Micro To Macro}
\arguments{
\item{micro}{vector of micro clusters ids to retrieve the assignment for}
}
\value{
microToMacro assignment
}
\description{
Assignment vector that associates each micro-cluster to a macro cluster
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
