% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_get_microclusters}
\alias{textClust_R_get_microclusters}
\title{get_microclusters}
\value{
list of micro-clusters
}
\description{
Getter for micro-clusters of textClust, returns interfaced C objects
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
