% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_get_microweights}
\alias{textClust_R_get_microweights}
\title{get_microweights}
\description{
Returns the weights of micro-clusters of an textClust_R object
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
