% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_getRepresentatives}
\alias{textClust_R_getRepresentatives}
\title{Representative Words for cluster}
\arguments{
\item{num}{number of words to return}

\item{type}{return representatives for micro or macro clusters}
}
\value{
a list containing the words for each micro-cluster
}
\description{
Returns the num words with the heighest weight for each micro-cluster
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
