% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_getSecondsPerMessage}
\alias{textClust_R_getSecondsPerMessage}
\title{Seconds per messages}
\description{
Returns the number of seconds used to process a message
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
}
