% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_getThroughputPerSecond}
\alias{textClust_R_getThroughputPerSecond}
\title{Returns Throughput}
\description{
Returns the number of processed messages per Second
}
\author{
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
