% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_serialize}
\alias{textClust_R_serialize}
\title{Serialize DSC_textClust_R object}
\description{
Exports the C++ Object of DSC_textClust_R to an R-List
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
