% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_updateMacroClusters}
\alias{textClust_R_updateMacroClusters}
\title{Update macro clusters}
\description{
Updates the macro clusters by running hierarchical clustering
}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
