% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_textClust.R
\name{DSC_textClust}
\alias{DSC_textClust}
\title{textClust}
\usage{
DSC_textClust(r = 0.4, lambda = 0.1, tgap = 1000, nmin = 1, nmax = 1,
  updateAll = F, k = NA_integer_, h = NA_real_, verbose = F,
  termFading = T)
}
\arguments{
\item{r}{distance threshold to merge two micro-clusters}

\item{lambda}{fading constant}

\item{tgap}{time between outlier removal}

\item{nmin}{min number of ngrams to use}

\item{nmax}{max number of ngrams to use}

\item{updateAll}{logical whether a new observations is used to update all micro-clusters within \code{r} or just the closest.}

\item{k}{number of clusters for macro-clustering}

\item{h}{height to determine number of clusters for macro-clustering}

\item{verbose}{logical whether to be more verbose}

\item{termFading}{logical whether individual terms should also be faded}
}
\value{
micro clusters
}
\description{
Stream Clustering algorithm which clusters messages. Contains a description string as well as a reference class object of \code{textClust} which is the work-horse.
}
\examples{
stream = DSD_ReadCSV("file.txt", sep = "\\t", comment.char="", quote="")
textClust <- DSC_textClust(r=.75, lambda=.1, tgap=1000)
update(textClust, stream, n=100)

}
\references{
Matthias Carnein, Dennis Assenmacher, Heike Trautmann (2017)"Stream Clustering of Chat Messages with Applications to Twitch Streams". In: Advances in Conceptual Modeling. ER 2017. Lecture Notes in Computer Science, vol 10651. Springer, Cham
}
\author{
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
