% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSC_textClust.R
\name{nclusters.DSC_textClust}
\alias{nclusters.DSC_textClust}
\title{Number of clusters for DSC_textClust}
\usage{
nclusters.DSC_textClust(x, type = c("auto", "micro", "macro"), ...)
}
\arguments{
\item{x}{object of class DSC_textClust}
}
\description{
Returns the number of clusters for DSC_textClust
}
\examples{
algorithm = DSC_textClust(lambda=0.001, r=0.2, tgap=100, verbose=T)
print(algorithm)

}
\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
