% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\docType{class}
\name{textClust_R-class}
\alias{textClust_R-class}
\alias{textClust_R}
\title{Reference Class textClust}
\description{
Maintains the current clustering and implements the stream package interfaces.
}
\section{Fields}{

\describe{
\item{\code{C}}{Exposed C++ class}

\item{\code{nmin}}{min number of ngrams to use}

\item{\code{nmax}}{max number of ngrams to use}

\item{\code{k}}{number of clusters}

\item{\code{h}}{height for hierarchical clustering}

\item{\code{microMacroAssignment}}{assignment vector associating micro-clusters to a maco-cluster}

\item{\code{throughput}}{stores the throughput of messages}

\item{\code{termFading}}{logical whether individual terms should be faded}

\item{\code{stopword}}{chracter vector of stopwords to remove}

\item{\code{linkage}}{method for hierarchical clustering}

\item{\code{weightedReclustering}}{logical whether reclustering should consider cluster weights}

\item{\code{minWeight}}{minimum weight of micro clusters to be used for reclustering}

\item{\code{textCol}}{index of column that contains the text}

\item{\code{timeCol}}{index of column that contains the timestamps}

\item{\code{timeFormat}}{string formatting of time Column}

\item{\code{timePrecision}}{Precision of fading, either seconds, minutes, hours or days}

\item{\code{fadeNaturalTime}}{Logical whether Natural Time or Number of observations should be used for fading}
}}


\author{
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
