% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textClust_R.R
\name{textClust_R_initialize}
\alias{textClust_R_initialize}
\title{Constructor of textClust}
\arguments{
\item{r}{radius threshold}

\item{lambda}{decay rate}

\item{tgap}{Time-interval for outlier detection and clean-up}

\item{nmin}{min number of ngrams to use}

\item{nmax}{max number of ngrams to use}

\item{updateAll}{logical whether a new observations is used to update all micro-clusters within \code{r} or just the closest.}

\item{k}{number of clusters}

\item{h}{height for hierarchical clustering}

\item{verbose}{logical whether to be more verbose}

\item{termFading}{logical whether individual terms should be faded}
}
\value{
refernce class object
}
\description{
Constructor of textClust
}
\author{
Dennis Assenmacher \email{dennis.assenmacher@wi.uni-muenster.de}
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
